requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
altair==5.5.0
altgraph==0.17.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==6.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.2
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
coverage==7.13.1
cryptg==0.3.1
cryptography==46.0.3
CTkMessagebox==2.7
customtkinter==5.2.2
darkdetect==0.8.0
deepface==0.0.96
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.120.4
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
frozenlist==1.7.0
gast==0.7.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hachoir==3.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
# Editable install with no version control (instabotpro-v2==0.0.0)
-e c:\users\naor1\onedrive\desktop\instabotpro_v2_clean_repo
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaitaistruct==0.11
keras==3.12.0
keyring==25.7.0
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
magic-filter==1.0.12
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.0.4
namex==0.1.0
narwhals==1.47.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pefile==2023.2.7
pillow==10.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.27.0
protobuf==6.31.1
psutil==7.2.1
psycopg2-binary==2.9.11
pyaes==1.6.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydivert==2.1.0
pyfiglet==0.8.post1
Pygments==2.19.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.11
PyJWT==2.10.1
pymongo==4.13.2
pyOpenSSL==25.3.0
pyparsing==3.3.1
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
pytesseract==0.3.13
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
replicate==1.0.7
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==12.6.0
rpds-py==0.26.0
rsa==4.9
ruff==0.14.14
selenium==4.39.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.49.1
streamlit==1.47.0
sympy==1.14.0
telebot==0.0.5
Telethon==1.36.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
tg-login==0.0.3
tgcf==0.5
toml==0.10.2
tomli==2.4.0
tornado==6.5.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
twilio==9.10.0
typer==0.12.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
verlat==0.1.0.post1
vulture==2.14
watchdog==6.0.0
watchfiles==1.1.1
watermark.py==0.0.3.post3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
